Move QuantumResourceService into separate class
