+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/services/QuantumResourceService.java
@@ -19,71 +19,32 @@
 
 package org.planqk.atlas.core.services;
 
-import java.util.NoSuchElementException;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
 
 import org.planqk.atlas.core.model.QuantumAlgorithm;
 import org.planqk.atlas.core.model.QuantumResource;
 import org.planqk.atlas.core.model.QuantumResourceType;
-import org.planqk.atlas.core.repository.AlgorithmRepository;
-import org.planqk.atlas.core.repository.QuantumResourceRepository;
-import org.planqk.atlas.core.repository.QuantumResourceTypeRepository;
 
-import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
 
-@Service
-@AllArgsConstructor
-public class QuantumResourceService {
+public interface QuantumResourceService {
+    @Transactional
+    void deleteQuantumResource(UUID resourceId);
 
-    private final AlgorithmService algorithmService;
-    private final AlgorithmRepository algorithmRepository;
-    private final QuantumResourceTypeRepository typeRepository;
-    private final QuantumResourceRepository resourceRepository;
+    Page<QuantumResourceType> findAllResourceTypes(Pageable pageable);
 
     @Transactional
-    public void deleteQuantumResource(UUID resourceId) {
-        resourceRepository.deleteById(resourceId);
-    }
-
-    public Page<QuantumResourceType> findAllResourceTypes(Pageable pageable) {
-        return typeRepository.findAll(pageable);
-    }
+    QuantumResourceType addOrUpdateQuantumResourceType(QuantumResourceType resourceType);
 
     @Transactional
-    public QuantumResourceType addOrUpdateQuantumResourceType(QuantumResourceType resourceType) {
-        return typeRepository.save(resourceType);
-    }
+    QuantumResource addOrUpdateQuantumResource(QuantumResource resource);
 
     @Transactional
-    public QuantumResource addOrUpdateQuantumResource(QuantumResource resource) {
-        if (resource.getQuantumResourceType().getId() == null) {
-            var type = addOrUpdateQuantumResourceType(resource.getQuantumResourceType());
-            resource.setQuantumResourceType(type);
-        }
-        return resourceRepository.save(resource);
-    }
+    QuantumResource addQuantumResourceToAlgorithm(QuantumAlgorithm algo, QuantumResource resource);
 
     @Transactional
-    public QuantumResource addQuantumResourceToAlgorithm(QuantumAlgorithm algo, QuantumResource resource) {
-        var updatedResource = resource;
-        if (updatedResource.getId() == null) {
-            updatedResource = this.addOrUpdateQuantumResource(resource);
-        }
-        updatedResource.setAlgorithm(algo);
-        return this.resourceRepository.save(updatedResource);
-    }
-
-    @Transactional
-    public QuantumResource addQuantumResourceToAlgorithm(UUID algoId, UUID resourceId) {
-        var resource = resourceRepository.findById(resourceId).orElseThrow(NoSuchElementException::new);
-        var algorithm = (QuantumAlgorithm) algorithmRepository.findById(algoId).orElseThrow(NoSuchElementException::new);
-
-        return addQuantumResourceToAlgorithm(algorithm, resource);
-    }
+    QuantumResource addQuantumResourceToAlgorithm(UUID algoId, UUID resourceId);
 }
